+++ NEWS.md
@@ -1,5 +1,8 @@
 # Version 0.1.1 [current]
 
+### Minor changes
+* Add `transform.attr` parameter to control parameter transformations
+
 ### Bugfixes
 * Specific errors about minority class with an unique instance
 * Change class to kknn pakcage in landmarking 1nn and eliteNN